Add files via upload
This project is written in 'c' language. In this work, I have tried to implement the research paper named "The Link Prediction Problem for Social Networks"  published on January 8, 2004. The link to get the research paper is https://onlinelibrary.wiley.com/doi/epdf/10.1002/asi.20591. It uses the undirected graph to achieve its objective.